@@ -4,15 +4,15 @@
   "version": "1.0.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
-    "cypress": "^3.1.0",
-    "cypress-image-snapshot": "^3.1.1",
-    "gatsby": "^2.0.118",
-    "gatsby-plugin-image": "^0.0.2",
-    "gatsby-plugin-sharp": "^2.0.20",
-    "gatsby-source-filesystem": "^2.3.35",
-    "gatsby-transformer-sharp": "^2.5.19",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0"
+    "cypress": "^7.3.0",
+    "cypress-image-snapshot": "^4.0.1",
+    "gatsby": "^3.6.0-next.5",
+    "gatsby-plugin-image": "^1.6.0-next.0",
+    "gatsby-plugin-sharp": "^3.6.0-next.0",
+    "gatsby-source-filesystem": "^3.6.0-next.0",
+    "gatsby-transformer-sharp": "^3.6.0-next.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   },
   "keywords": [
     "gatsby"
@@ -33,10 +33,10 @@
   "devDependencies": {
     "cross-env": "^5.2.0",
     "cypress-junit-reporter": "^1.3.1",
-    "gatsby-cypress": "0.4.11",
+    "gatsby-cypress": "1.6.0-next.0",
     "is-ci": "^2.0.0",
     "prettier": "2.0.4",
-    "start-server-and-test": "^1.7.1"
+    "start-server-and-test": "^1.12.2"
   },
   "repository": {
     "type": "git",
