@@ -5,12 +5,12 @@
   "version": "1.0.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
-    "core-js": "^2.6.11",
     "gatsby": "next",
     "gatsby-core-utils": "next",
     "gatsby-image": "next",
     "gatsby-plugin-glamor": "next",
     "gatsby-plugin-google-analytics": "next",
+    "gatsby-plugin-image": "^1.5.0",
     "gatsby-plugin-manifest": "next",
     "gatsby-plugin-offline": "next",
     "gatsby-plugin-sharp": "next",
@@ -25,15 +25,14 @@
     "mousetrap": "^1.6.5",
     "progress": "^2.0.3",
     "prop-types": "^15.7.2",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-gravatar": "^2.6.3",
-    "react-icons": "^2.2.7",
+    "react-icons": "2",
     "react-modal": "^3.11.2",
     "react-typography": "^0.16.19",
     "request": "^2.88.2",
-    "slug": "^0.9.4",
-    "typeface-space-mono": "0.0.71",
+    "typeface-space-mono": "1.1.13",
     "typography": "^0.16.19"
   },
   "keywords": [
