@@ -17,11 +17,10 @@
     "glamor": "^2.20.40",
     "gray-percentage": "^2.0.0",
     "lodash": "^4.17.20",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-icons": "^2.2.7",
     "react-typography": "^0.16.19",
-    "slash": "^1.0.0",
     "typeface-josefin-sans": "^0.0.75",
     "typeface-josefin-slab": "^0.0.72",
     "typeface-rochester": "^0.0.72",
