@@ -1,9 +1,10 @@
 module.exports = {
   flags: {
     FUNCTIONS: true,
+    DEV_SSR: true,
   },
   siteMetadata: {
     title: "functions",
   },
-  plugins: [`gatsby-plugin-gatsby-cloud`],
+  plugins: [`gatsby-plugin-gatsby-cloud`, `gatsby-plugin-cool`],
 }
