@@ -6,6 +6,8 @@ const {
   default: fetchGraphql,
 } = require("gatsby-source-wordpress/dist/utils/fetch-graphql")
 
+const gatsbyConfig = require("../gatsby-config")
+
 const { testResolvedData } = require("./test-utils/test-resolved-data")
 const { queries } = require("./test-utils/queries")
 
@@ -16,6 +18,11 @@ jest.setTimeout(100000)
 const isWarmCache = process.env.WARM_CACHE
 const url = `http://localhost:8000/___graphql`
 
+const getPluginConfig = () =>
+  gatsbyConfig.plugins.find(
+    plugin => typeof plugin === 'object' && plugin.resolve === `gatsby-source-wordpress`
+  )
+
 describe(`data resolution`, () => {
   it(`resolves correct number of nodes`, async () => {
     const { data } = await fetchGraphql({
@@ -25,11 +32,11 @@ describe(`data resolution`, () => {
 
     expect(data[`allWpMediaItem`].nodes).toBeTruthy()
     expect(data[`allWpMediaItem`].nodes).toMatchSnapshot()
-    expect(data[`allWpMediaItem`].totalCount).toBe(15)
+    expect(data[`allWpMediaItem`].totalCount).toBe(17)
 
     expect(data[`allWpTag`].totalCount).toBe(5)
     expect(data[`allWpUser`].totalCount).toBe(1)
-    expect(data[`allWpPage`].totalCount).toBe(4)
+    expect(data[`allWpPage`].totalCount).toBe(5)
     expect(data[`allWpPost`].totalCount).toBe(5)
     expect(data[`allWpComment`].totalCount).toBe(1)
     expect(data[`allWpTaxonomy`].totalCount).toBe(3)
@@ -78,7 +85,7 @@ describe(`data resolution`, () => {
     })
 
     expect(gatsbyResult.data.allWpTermNode.nodes.length).toBe(14)
-    expect(gatsbyResult.data.allWpContentNode.nodes.length).toBe(27)
+    expect(gatsbyResult.data.allWpContentNode.nodes.length).toBe(30)
   })
 
   it(`resolves interface fields which are a mix of Gatsby nodes and regular object data with no node`, async () => {
@@ -307,5 +314,133 @@ describe(`data resolution`, () => {
 
     expect(gatsbyResult.data.wpPage).toStrictEqual(wpGraphQLPageNormalizedPaths)
     expect(gatsbyResult.data.wp.seo).toStrictEqual(WPGraphQLData.seo)
+  })
+
+  it(`Does not download files whose size exceed the maxFileSizeBytes option`, async () => {
+    const wpPluginOpts = getPluginConfig()
+    const { maxFileSizeBytes } = wpPluginOpts.options.type.MediaItem.localFile
+    /**
+     * Ensure that the fileSize "gt" filter value matches the maxFileSizeBytes value in gatsby-config
+     */
+    const { data: { allWpMediaItem: { nodes }}} = await fetchGraphql({
+      url,
+      query: /* GraphQL */`
+        query tooLargeFiles($maxFileSizeBytes: Int!, $includedMimeTypes: [String]!) {
+          allWpMediaItem(filter: { fileSize: { gt: $maxFileSizeBytes }, mimeType: {in: $includedMimeTypes } }) {
+            nodes {
+              id
+              sourceUrl
+              fileSize
+              localFile {
+                absolutePath
+                size
+              }
+            }
+          }
+        } 
+      `,
+      variables: {
+        maxFileSizeBytes,
+        includedMimeTypes: ['image/jpeg'],
+      }
+    })
+
+    expect(nodes.length).toEqual(1)
+    nodes.forEach(node => {
+      expect(node.localFile).toEqual(null)
+    })
+  })
+
+  it(`Downloads files and creates a local file node for files whose size are <= maxFileSizeBytes`, async () => {
+    const wpPluginOpts = getPluginConfig()
+    const { maxFileSizeBytes } = wpPluginOpts.options.type.MediaItem.localFile
+    /**
+     * Ensure that the fileSize "gt" filter value matches the maxFileSizeBytes value in gatsby-config
+     */
+    const { data: { allWpMediaItem: { nodes }}} = await fetchGraphql({
+      url,
+      query: /* GraphQL */`
+        query tooLargeFiles($maxFileSizeBytes: Int!, $includedMimeTypes: [String]!) {
+          allWpMediaItem(filter: { fileSize: { lte: $maxFileSizeBytes }, mimeType: {in: $includedMimeTypes } }) {
+            nodes {
+              id
+              sourceUrl
+              fileSize
+              localFile {
+                absolutePath
+                size
+              }
+            }
+          }
+        } 
+      `,
+      variables: {
+        maxFileSizeBytes,
+        includedMimeTypes: ['image/jpeg'],
+      }
+    })
+
+    nodes.forEach(node => {
+      expect(node.localFile).toBeDefined()
+    })
+  })
+
+  it(`Does not download files whose mime types are excluded`, async () => {
+    const wpPluginOpts = getPluginConfig()
+    const { excludeByMimeTypes } = wpPluginOpts.options.type.MediaItem.localFile
+
+    const { data: { allWpMediaItem: { nodes }}} = await fetchGraphql({
+      url,
+      query: /* GraphQL */`
+        query excludedMimeType($excludeByMimeTypes: [String]) {
+          allWpMediaItem(filter: { mimeType: { in: $excludeByMimeTypes }}) {
+            nodes {
+              id
+              mimeType
+              localFile {
+                size
+              }
+            }
+          }
+        }
+      `,
+      variables: {
+        excludeByMimeTypes,
+      },
+    })
+
+    expect(nodes.length).toEqual(2)
+    nodes.forEach(node => {
+      expect(node.localFile).toEqual(null)
+    })
+  })
+
+  it(`Creats a local file node for files not excluded by the "excludedByMimeTypes" option`, async () => {
+    const wpPluginOpts = getPluginConfig()
+    const { excludeByMimeTypes } = wpPluginOpts.options.type.MediaItem.localFile
+
+    const { data: { allWpMediaItem: { nodes }}} = await fetchGraphql({
+      url,
+      query: /* GraphQL */`
+        query excludedMimeType($excludeByMimeTypes: [String]) {
+          allWpMediaItem(filter: { mimeType: { nin: $excludeByMimeTypes }}) {
+            nodes {
+              id
+              mimeType
+              localFile {
+                size
+              }
+            }
+          }
+        }
+      `,
+      variables: {
+        excludeByMimeTypes,
+      },
+    })
+
+    nodes.forEach(node => {
+      expect(node.localFile).toBeDefined()
+    })
   })
 })
