@@ -21,10 +21,10 @@ exports.gatsbyCleanBeforeAll = async done => {
   }
 }
 
-exports.spawnGatsbyProcess = (command = `develop`, env = {}) =>
-  spawn(
+exports.spawnGatsbyProcess = (command = `develop`, env = {}) => {
+  const proc = spawn(
     process.execPath,
-    [gatsbyBin, command, ...(command === `develop` ? ["-H", "localhost"] : [])],
+    [gatsbyBin, command, ...(command === `develop` ? ["-H", "localhost", "--port", "8000"] : [])],
     {
       stdio: [`inherit`, `inherit`, `inherit`],
       env: {
@@ -34,3 +34,8 @@ exports.spawnGatsbyProcess = (command = `develop`, env = {}) =>
       },
     }
   )
+
+  process.on('SIGINT', proc.kill)
+
+  return proc
+}
