@@ -141,9 +141,9 @@ function buildLocalCommands(cli: yargs.Argv, isLocalSite: boolean): void {
       _.option(`H`, {
         alias: `host`,
         type: `string`,
-        default: process.env.HOST || defaultHost,
-        describe: process.env.HOST
-          ? `Set host. Defaults to ${process.env.HOST} (set by env.HOST) (otherwise defaults ${defaultHost})`
+        default: process.env.GATSBY_HOST || defaultHost,
+        describe: process.env.GATSBY_HOST
+          ? `Set host. Defaults to ${process.env.GATSBY_HOST} (set by env.GATSBY_HOST) (otherwise defaults ${defaultHost})`
           : `Set host. Defaults to ${defaultHost}`,
       })
         .option(`p`, {
@@ -203,8 +203,11 @@ function buildLocalCommands(cli: yargs.Argv, isLocalSite: boolean): void {
         `develop`,
         (args: yargs.Arguments, cmd: (args: yargs.Arguments) => unknown) => {
           process.env.NODE_ENV = process.env.NODE_ENV || `development`
-          const { startGraphQLServer } = require(`gatsby-recipes`)
-          startGraphQLServer(siteInfo.directory, true)
+
+          if (process.env.GATSBY_EXPERIMENTAL_ENABLE_ADMIN) {
+            const { startGraphQLServer } = require(`gatsby-recipes`)
+            startGraphQLServer(siteInfo.directory, true)
+          }
 
           if (args.hasOwnProperty(`inspect`)) {
             args.inspect = args.inspect || 9229
