@@ -1,41 +1,92 @@
 import React from "react"
-import "@testing-library/jest-dom"
+import "@testing-library/jest-dom/extend-expect"
 import userEvent from "@testing-library/user-event"
-import { render, screen } from "@testing-library/react"
+import { render, screen, act, waitFor } from "@testing-library/react"
 
-import { wrapPageElement } from "../gatsby-browser"
-import Indicator from "../indicator"
+import { wrapRootElement } from "../gatsby-browser"
+import Indicator from "../components/Indicator"
+
+import { server } from "./mocks/server"
+
+const createUrl = path => `https://test.com/${path}`
+const copyLinkMessage = `Copy link`
+const infoButtonMessage = `Preview updated`
+const { location } = window
+
+process.env.GATSBY_PREVIEW_AUTH_TOKEN = `token`
 
 describe(`Preview status indicator`, () => {
-  const waitForPoll = ms =>
-    new Promise(resolve => setTimeout(resolve, ms || 50))
+  const assertTooltipText = async ({ route, text, matcherType }) => {
+    process.env.GATSBY_PREVIEW_API_URL = createUrl(route)
 
-  describe(`wrapPageElement`, () => {
+    // it will disable setTimeout behaviour - only fetchData once
+    jest.useFakeTimers()
+
+    await act(async () => {
+      render(<Indicator />)
+    })
+
+    if (matcherType === `query`) {
+      await waitFor(() => {
+        expect(
+          screen.queryByText(text, { exact: false })
+        ).not.toBeInTheDocument()
+      })
+    } else if (matcherType === `get`) {
+      await waitFor(() => {
+        expect(screen.getByText(text, { exact: false })).toBeInTheDocument()
+      })
+    }
+  }
+
+  beforeAll(() => {
+    server.listen()
+  })
+
+  beforeEach(() => {
+    // reset all mocks
+    jest.resetModules()
+    global.fetch = require(`node-fetch`)
+    jest.spyOn(global, `fetch`)
+    // delete window.location
+    //   window.location = {
+    //     href: `https://preview-xxx.gtsb.io`,
+    //     host: `preview-xxx.gtsb.io`,
+    //     hostname: `preview-xxx.gtsb.io`,
+    //     reload: jest.fn(),
+    //     replace: jest.fn()
+    //   }
+  })
+
+  afterEach(() => {
+    jest.useRealTimers()
+    server.resetHandlers()
+    // window.location = location
+  })
+
+  afterAll(() => {
+    server.close()
+  })
+
+  describe(`wrapRootElement`, () => {
     const testMessage = `Test Page`
 
     beforeEach(() => {
-      global.fetch = jest.fn(() =>
-        Promise.resolve({
-          json: () => {
-            return {
-              currentBuild: { id: `123`, buildStatus: `SUCCESS` },
-              latestBuild: { id: `1234`, buildStatus: `SUCCESS` },
-            }
-          },
-        })
-      )
+      process.env.GATSBY_PREVIEW_API_URL = createUrl(`success`)
     })
 
     it(`renders the initial page and indicator if indicator enabled`, async () => {
+      // do not fetch any data
+      global.fetch = jest.fn(() => new Promise(() => {}))
       process.env.GATSBY_PREVIEW_INDICATOR_ENABLED = `true`
 
-      render(
-        wrapPageElement({
-          element: <div>{testMessage}</div>,
-        })
-      )
-
-      await waitForPoll()
+      act(() => {
+        render(
+          wrapRootElement({
+            element: <div>{testMessage}</div>,
+          })
+        )
+      })
 
       expect(screen.getByText(testMessage)).toBeInTheDocument()
       expect(
@@ -47,7 +98,7 @@ describe(`Preview status indicator`, () => {
       process.env.GATSBY_PREVIEW_INDICATOR_ENABLED = `false`
 
       render(
-        wrapPageElement({
+        wrapRootElement({
           element: <div>{testMessage}</div>,
         })
       )
@@ -58,9 +109,9 @@ describe(`Preview status indicator`, () => {
       ).not.toBeInTheDocument()
     })
 
-    it(`renders initial page without indicator if Indicator errors`, async () => {
+    it(`renders initial page without indicator if api errors`, async () => {
       render(
-        wrapPageElement({
+        wrapRootElement({
           element: <div>{testMessage}</div>,
         })
       )
@@ -69,8 +120,6 @@ describe(`Preview status indicator`, () => {
         Promise.resolve({ json: () => new Error(`failed`) })
       )
 
-      await waitForPoll()
-
       expect(screen.getByText(testMessage)).toBeInTheDocument()
       expect(
         screen.queryByTestId(`preview-status-indicator`)
@@ -78,97 +127,175 @@ describe(`Preview status indicator`, () => {
     })
   })
 
-  describe(`Indicator component`, () => {
-    beforeEach(() => {
-      render(<Indicator />)
-    })
+  describe(`Indicator`, () => {
+    describe(`Gatsby Button`, () => {
+      it(`should show a more recent succesful build when available`, async () => {
+        await assertTooltipText({
+          route: `success`,
+          text: `New preview available`,
+          matcherType: `get`,
+        })
+      })
 
-    describe(`Success state`, () => {
-      beforeEach(async () => {
-        global.fetch = jest.fn(() =>
-          Promise.resolve({
-            json: () => {
-              return {
-                currentBuild: { id: `123`, buildStatus: `SUCCESS` },
-                latestBuild: { id: `1234`, buildStatus: `SUCCESS` },
-              }
-            },
-          })
+      it(`should show an error message when most recent build fails`, async () => {
+        await assertTooltipText({
+          route: `error`,
+          text: `Preview error`,
+          matcherType: `get`,
+        })
+      })
+
+      it(`should show a preview building message when most recent build is building`, async () => {
+        await assertTooltipText({
+          route: `building`,
+          text: `Building a new preview`,
+          matcherType: `get`,
+        })
+      })
+
+      it(`should have loading state if no build data has been fetched`, async () => {
+        await assertTooltipText({
+          route: `fetching`,
+          text: `Fetching preview info...`,
+          matcherType: `get`,
+        })
+      })
+
+      it(`should have no tooltip when preview is up to date`, async () => {
+        await assertTooltipText({
+          route: `uptodate`,
+          text: `Fetching preview info...`,
+          matcherType: `query`,
+        })
+      })
+
+      it(`should open a new window to build logs when tooltip is clicked on error`, async () => {
+        process.env.GATSBY_PREVIEW_API_URL = createUrl(`error`)
+        window.open = jest.fn()
+
+        let gatsbyButtonTooltip
+        const pathToBuildLogs = `https://www.gatsbyjs.com/dashboard/999/sites/111/builds/123/details`
+        const returnTo = encodeURIComponent(pathToBuildLogs)
+
+        await act(async () => {
+          render(<Indicator />)
+        })
+
+        await waitFor(() => {
+          gatsbyButtonTooltip = screen.getByText(`View logs`, { exact: false })
+        })
+
+        userEvent.click(gatsbyButtonTooltip)
+        expect(window.open).toHaveBeenCalledWith(
+          `${pathToBuildLogs}?returnTo=${returnTo}`
         )
+      })
+    })
 
-        await waitForPoll()
+    describe(`Link Button`, () => {
+      // TODO: Test hoverability
+
+      it(`should have no tooltip when successful`, async () => {
+        await assertTooltipText({
+          route: `success`,
+          text: copyLinkMessage,
+          matcherType: `query`,
+        })
       })
 
-      it(`renders when more recent successful build available`, async () => {
-        expect(screen.getByText(`New preview available`)).toBeInTheDocument()
+      it(`should have no tooltip when error`, async () => {
+        await assertTooltipText({
+          route: `error`,
+          text: copyLinkMessage,
+          matcherType: `query`,
+        })
       })
 
-      it(`navigates to new build when indicator is clicked`, async () => {
-        delete window.location
-        window.location = new URL(`https://preview-testsite.gtsb.io`)
-        window.location.replace = jest.fn(
-          () => (window.location = new URL(`https://build-123.gtsb.io`))
+      it(`should have a copy link tooltip when building`, async () => {
+        await assertTooltipText({
+          route: `building`,
+          text: copyLinkMessage,
+          matcherType: `get`,
+        })
+      })
+
+      it(`should have a copy link tooltip in initial state`, async () => {
+        await assertTooltipText({
+          route: `fetching`,
+          text: copyLinkMessage,
+          matcherType: `get`,
+        })
+      })
+
+      it(`should have a copy link tooltip when up to date`, async () => {
+        await assertTooltipText({
+          route: `uptodate`,
+          text: copyLinkMessage,
+          matcherType: `get`,
+        })
+      })
+
+      it(`should copy to clipboard when copy link is clicked`, async () => {
+        process.env.GATSBY_PREVIEW_API_URL = createUrl(`uptodate`)
+
+        navigator.clipboard = { writeText: jest.fn() }
+        let copyLinkButton
+
+        await act(async () => {
+          render(<Indicator />)
+        })
+
+        await waitFor(() => {
+          copyLinkButton = screen.getByText(copyLinkMessage, { exact: false })
+        })
+
+        userEvent.click(copyLinkButton)
+        expect(navigator.clipboard.writeText).toHaveBeenCalledWith(
+          `http://localhost/`
         )
-
-        const previewIndicator = screen.getByText(`New preview available`)
-        userEvent.click(previewIndicator)
-        await waitForPoll(300)
-
-        expect(String(window.location)).toBe(`https://build-123.gtsb.io/`)
       })
     })
 
-    it(`renders FAILED state when most recent build failed`, async () => {
-      global.fetch = jest.fn(() =>
-        Promise.resolve({
-          json: () => {
-            return {
-              currentBuild: { id: `123`, buildStatus: `ERROR` },
-              latestBuild: { id: `1234`, buildStatus: `SUCCESS` },
-            }
-          },
+    describe(`Info Button`, () => {
+      it(`should have no tooltip when successful`, async () => {
+        await assertTooltipText({
+          route: `success`,
+          text: infoButtonMessage,
+          matcherType: `query`,
         })
-      )
+      })
 
-      await waitForPoll()
-
-      expect(
-        screen.getByText(`Latest preview build failed`)
-      ).toBeInTheDocument()
-    })
-
-    it(`renders BUILDING state when most recent build is currently building`, async () => {
-      global.fetch = jest.fn(() =>
-        Promise.resolve({
-          json: () => {
-            return {
-              currentBuild: { id: `123`, buildStatus: `BUILDING` },
-              latestBuild: { id: `1234`, buildStatus: `SUCCESS` },
-            }
-          },
+      it(`should have no tooltip when error`, async () => {
+        await assertTooltipText({
+          route: `error`,
+          text: infoButtonMessage,
+          matcherType: `query`,
         })
-      )
+      })
 
-      await waitForPoll()
-
-      expect(screen.getByText(`New preview building`)).toBeInTheDocument()
-    })
-
-    it(`renders NO state when on most successful build`, async () => {
-      global.fetch = jest.fn(() =>
-        Promise.resolve({
-          json: () => {
-            return {
-              currentBuild: { id: `123`, buildStatus: `SUCCESS` },
-              latestBuild: { id: `123`, buildStatus: `SUCCESS` },
-            }
-          },
+      it(`should have no tooltip when building`, async () => {
+        await assertTooltipText({
+          route: `building`,
+          text: infoButtonMessage,
+          matcherType: `query`,
         })
-      )
+      })
 
-      expect(
-        screen.queryByTestId(`preview-status-indicator`)
-      ).not.toBeInTheDocument()
+      it(`should have no tooltip in initial state`, async () => {
+        await assertTooltipText({
+          route: `fetching`,
+          text: infoButtonMessage,
+          matcherType: `query`,
+        })
+      })
+
+      it(`should have a last updated tooltip when up to date`, async () => {
+        await assertTooltipText({
+          route: `uptodate`,
+          text: infoButtonMessage,
+          matcherType: `get`,
+        })
+      })
     })
   })
 })
