@@ -273,9 +273,19 @@ const applyCachingHeaders = (
     return headers
   }
 
-  const chunks = Array.from(pluginData.pages.values()).map(
-    page => page.componentChunkName
-  )
+  let chunks = []
+  // Gatsby v3.5 added componentChunkName to store().components
+  // So we prefer to pull chunk names off that as it gets very expensive to loop
+  // over large numbers of pages.
+  const isComponentChunkSet = !!pluginData.components.entries()?.next()
+    ?.value[1]?.componentChunkName
+  if (isComponentChunkSet) {
+    chunks = [...pluginData.components.values()].map(c => c.componentChunkName)
+  } else {
+    chunks = Array.from(pluginData.pages.values()).map(
+      page => page.componentChunkName
+    )
+  }
 
   chunks.push(`pages-manifest`, `app`)
 
