@@ -1,9 +1,32 @@
-import React from "react"
-import Indicator from "./indicator"
+import React, { useEffect, useState } from "react"
+import { createPortal } from "react-dom"
+import Indicator from "./components/Indicator"
 
-export const wrapPageElement = ({ element }) => {
+function PreviewIndicatorRoot() {
+  const [indicatorRootRef, setIndicatorRootRef] = useState()
+
+  useEffect(() => {
+    const indicatorRoot = document.createElement(`div`)
+    indicatorRoot.id = `gatsby-preview-indicator`
+    setIndicatorRootRef(indicatorRoot)
+    document.body.appendChild(indicatorRoot)
+  }, [])
+
+  if (!indicatorRootRef) {
+    return null
+  }
+
+  return createPortal(<Indicator />, indicatorRootRef)
+}
+
+export const wrapRootElement = ({ element }) => {
   if (process.env.GATSBY_PREVIEW_INDICATOR_ENABLED === `true`) {
-    return <Indicator>{element}</Indicator>
+    return (
+      <>
+        {element}
+        <PreviewIndicatorRoot />
+      </>
+    )
   } else {
     return <>{element}</>
   }
