@@ -8,7 +8,7 @@ export function buildPrefixer(prefix, ...paths) {
 // shape of `static-entry.js`. With it, we can build headers that point to the correct
 // hashed filenames and ensure we pull in the componentChunkName.
 export default function makePluginData(store, assetsManifest, pathPrefix) {
-  const { program, pages: storePages } = store.getState()
+  const { program, pages, components } = store.getState()
   const publicFolder = buildPrefixer(program.directory, `public`)
   const functionsFolder = buildPrefixer(
     program.directory,
@@ -18,13 +18,13 @@ export default function makePluginData(store, assetsManifest, pathPrefix) {
   const stats = require(publicFolder(`webpack.stats.json`))
   // Get all the files, not just the first
   const chunkManifest = stats.assetsByChunkName
-  const pages = storePages
 
   // We combine the manifest of JS and the manifest of assets to make a lookup table.
   const manifest = { ...assetsManifest, ...chunkManifest }
 
   return {
     pages,
+    components,
     manifest,
     program,
     pathPrefix,
