@@ -3,8 +3,7 @@ import { mapValues, isPlainObject, trim } from "lodash"
 import webpack from "webpack"
 import HtmlWebpackPlugin from "html-webpack-plugin"
 import MiniCssExtractPlugin from "mini-css-extract-plugin"
-// TODO: swap back when https://github.com/geowarin/friendly-errors-webpack-plugin/pull/86 lands
-import FriendlyErrorsPlugin from "@pieh/friendly-errors-webpack-plugin"
+import FriendlyErrorsPlugin from "@soda/friendly-errors-webpack-plugin"
 import CopyPlugin from "copy-webpack-plugin"
 import HtmlWebpackTagsPlugin from "html-webpack-tags-plugin"
 
