@@ -78,27 +78,27 @@ describe(`gatsby-source-drupal`, () => {
 
   it(`Generates nodes`, () => {
     expect(Object.keys(nodes).length).not.toEqual(0)
-    expect(nodes[createNodeId(`file-1`)]).toBeDefined()
-    expect(nodes[createNodeId(`file-2`)]).toBeDefined()
-    expect(nodes[createNodeId(`tag-1`)]).toBeDefined()
-    expect(nodes[createNodeId(`tag-2`)]).toBeDefined()
-    expect(nodes[createNodeId(`article-1`)]).toBeDefined()
-    expect(nodes[createNodeId(`article-2`)]).toBeDefined()
-    expect(nodes[createNodeId(`article-3`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.file-1`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.file-2`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.tag-1`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.tag-2`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.article-1`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.article-2`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.article-3`)]).toBeDefined()
   })
 
   it(`Nodes contain contentDigest`, () => {
-    expect(nodes[createNodeId(`file-1`)]).toEqual(
+    expect(nodes[createNodeId(`und.file-1`)]).toEqual(
       objectContaining({
         internal: objectContaining({ contentDigest: `contentDigest` }),
       })
     )
-    expect(nodes[createNodeId(`article-2`)]).toEqual(
+    expect(nodes[createNodeId(`und.article-2`)]).toEqual(
       objectContaining({
         internal: objectContaining({ contentDigest: `contentDigest` }),
       })
     )
-    expect(nodes[createNodeId(`tag-1`)]).toEqual(
+    expect(nodes[createNodeId(`und.tag-1`)]).toEqual(
       objectContaining({
         internal: objectContaining({ contentDigest: `contentDigest` }),
       })
@@ -106,61 +106,64 @@ describe(`gatsby-source-drupal`, () => {
   })
 
   it(`Nodes contain attributes data`, () => {
-    expect(nodes[createNodeId(`file-1`)].filename).toEqual(`main-image.png`)
-    expect(nodes[createNodeId(`article-2`)].title).toEqual(`Article #2`)
-    expect(nodes[createNodeId(`tag-1`)].langcode).toEqual(`en`)
+    expect(nodes[createNodeId(`und.file-1`)].filename).toEqual(`main-image.png`)
+    expect(nodes[createNodeId(`und.article-2`)].title).toEqual(`Article #2`)
+    expect(nodes[createNodeId(`und.tag-1`)].langcode).toEqual(`en`)
   })
 
   it(`Preserves attributes.id`, () => {
-    expect(nodes[createNodeId(`article-2`)]._attributes_id).toEqual(22)
+    expect(nodes[createNodeId(`und.article-2`)]._attributes_id).toEqual(22)
   })
 
   it(`Handles 1:1 relationship`, () => {
     expect(
-      nodes[createNodeId(`article-1`)].relationships.field_main_image___NODE
+      nodes[createNodeId(`und.article-1`)].relationships.field_main_image___NODE
     ).not.toBeDefined()
     expect(
-      nodes[createNodeId(`article-2`)].relationships.field_main_image___NODE
-    ).toEqual(createNodeId(`file-1`))
+      nodes[createNodeId(`und.article-2`)].relationships.field_main_image___NODE
+    ).toEqual(createNodeId(`und.file-1`))
     expect(
-      nodes[createNodeId(`article-3`)].relationships.field_main_image___NODE
-    ).toEqual(createNodeId(`file-1`))
+      nodes[createNodeId(`und.article-3`)].relationships.field_main_image___NODE
+    ).toEqual(createNodeId(`und.file-1`))
   })
 
   it(`Handles 1:N relationship`, () => {
     expect(
-      nodes[createNodeId(`article-1`)].relationships.field_tags___NODE
+      nodes[createNodeId(`und.article-1`)].relationships.field_tags___NODE
     ).toEqual(
-      expect.arrayContaining([createNodeId(`tag-1`), createNodeId(`tag-2`)])
+      expect.arrayContaining([
+        createNodeId(`und.tag-1`),
+        createNodeId(`und.tag-2`),
+      ])
     )
     expect(
-      nodes[createNodeId(`article-2`)].relationships.field_tags___NODE
+      nodes[createNodeId(`und.article-2`)].relationships.field_tags___NODE
     ).not.toBeDefined()
     expect(
-      nodes[createNodeId(`article-3`)].relationships.field_tags___NODE
-    ).toEqual(expect.arrayContaining([createNodeId(`tag-1`)]))
+      nodes[createNodeId(`und.article-3`)].relationships.field_tags___NODE
+    ).toEqual(expect.arrayContaining([createNodeId(`und.tag-1`)]))
   })
 
   it(`Creates back references`, () => {
     expect(
-      nodes[createNodeId(`file-1`)].relationships[`node__article___NODE`]
+      nodes[createNodeId(`und.file-1`)].relationships[`node__article___NODE`]
     ).toEqual(
       expect.arrayContaining([
-        createNodeId(`article-2`),
-        createNodeId(`article-3`),
+        createNodeId(`und.article-2`),
+        createNodeId(`und.article-3`),
       ])
     )
     expect(
-      nodes[createNodeId(`tag-1`)].relationships[`node__article___NODE`]
+      nodes[createNodeId(`und.tag-1`)].relationships[`node__article___NODE`]
     ).toEqual(
       expect.arrayContaining([
-        createNodeId(`article-1`),
-        createNodeId(`article-3`),
+        createNodeId(`und.article-1`),
+        createNodeId(`und.article-3`),
       ])
     )
     expect(
-      nodes[createNodeId(`tag-2`)].relationships[`node__article___NODE`]
-    ).toEqual(expect.arrayContaining([createNodeId(`article-1`)]))
+      nodes[createNodeId(`und.tag-2`)].relationships[`node__article___NODE`]
+    ).toEqual(expect.arrayContaining([createNodeId(`und.article-1`)]))
   })
 
   it(`Download files without Basic Auth`, () => {
@@ -264,24 +267,30 @@ describe(`gatsby-source-drupal`, () => {
     describe(`Update content`, () => {
       describe(`Before update`, () => {
         it(`Attributes`, () => {
-          expect(nodes[createNodeId(`article-3`)].title).toBe(`Article #3`)
+          expect(nodes[createNodeId(`und.article-3`)].title).toBe(`Article #3`)
         })
         it(`Relationships`, () => {
-          expect(nodes[createNodeId(`article-3`)].relationships).toEqual({
-            field_main_image___NODE: createNodeId(`file-1`),
-            field_tags___NODE: [createNodeId(`tag-1`)],
+          expect(nodes[createNodeId(`und.article-3`)].relationships).toEqual({
+            field_main_image___NODE: createNodeId(`und.file-1`),
+            field_tags___NODE: [createNodeId(`und.tag-1`)],
           })
         })
         it(`Back references`, () => {
           expect(
-            nodes[createNodeId(`file-1`)].relationships[`node__article___NODE`]
-          ).toContain(createNodeId(`article-3`))
+            nodes[createNodeId(`und.file-1`)].relationships[
+              `node__article___NODE`
+            ]
+          ).toContain(createNodeId(`und.article-3`))
           expect(
-            nodes[createNodeId(`tag-1`)].relationships[`node__article___NODE`]
-          ).toContain(createNodeId(`article-3`))
+            nodes[createNodeId(`und.tag-1`)].relationships[
+              `node__article___NODE`
+            ]
+          ).toContain(createNodeId(`und.article-3`))
           expect(
-            nodes[createNodeId(`tag-2`)].relationships[`node__article___NODE`]
-          ).not.toContain(createNodeId(`article-3`))
+            nodes[createNodeId(`und.tag-2`)].relationships[
+              `node__article___NODE`
+            ]
+          ).not.toContain(createNodeId(`und.article-3`))
         })
       })
 
@@ -295,39 +304,45 @@ describe(`gatsby-source-drupal`, () => {
           })
         })
         it(`Attributes`, () => {
-          expect(nodes[createNodeId(`article-3`)].title).toBe(
+          expect(nodes[createNodeId(`und.article-3`)].title).toBe(
             `Article #3 - Updated`
           )
         })
         it(`Relationships`, () => {
           // removed `field_main_image`, changed `field_tags`
-          expect(nodes[createNodeId(`article-3`)].relationships).toEqual({
-            field_tags___NODE: [createNodeId(`tag-2`)],
+          expect(nodes[createNodeId(`und.article-3`)].relationships).toEqual({
+            field_tags___NODE: [createNodeId(`und.tag-2`)],
           })
         })
         it(`Back references`, () => {
           // removed `field_main_image`, `file-1` no longer has back reference to `article-3`
           expect(
-            nodes[createNodeId(`file-1`)].relationships[`node__article___NODE`]
-          ).not.toContain(createNodeId(`article-3`))
+            nodes[createNodeId(`und.file-1`)].relationships[
+              `node__article___NODE`
+            ]
+          ).not.toContain(createNodeId(`und.article-3`))
           // changed `field_tags`, `tag-1` no longer has back reference to `article-3`
           expect(
-            nodes[createNodeId(`tag-1`)].relationships[`node__article___NODE`]
-          ).not.toContain(createNodeId(`article-3`))
+            nodes[createNodeId(`und.tag-1`)].relationships[
+              `node__article___NODE`
+            ]
+          ).not.toContain(createNodeId(`und.article-3`))
           // changed `field_tags`, `tag-2` now has back reference to `article-3`
           expect(
-            nodes[createNodeId(`tag-2`)].relationships[`node__article___NODE`]
-          ).toContain(createNodeId(`article-3`))
+            nodes[createNodeId(`und.tag-2`)].relationships[
+              `node__article___NODE`
+            ]
+          ).toContain(createNodeId(`und.article-3`))
         })
       })
     })
 
     describe(`Insert content`, () => {
       it(`Node doesn't exist before webhook`, () => {
-        expect(nodes[createNodeId(`article-4`)]).not.toBeDefined()
+        expect(nodes[createNodeId(`und.article-4`)]).not.toBeDefined()
         expect(
-          nodes[createNodeId(`tag-1`)].relationships[`node__article___NODE`]
-        ).not.toContain(createNodeId(`article-4`))
+          nodes[createNodeId(`und.tag-1`)].relationships[`node__article___NODE`]
+        ).not.toContain(createNodeId(`und.article-4`))
       })
 
       describe(`After insert`, () => {
@@ -340,14 +355,16 @@ describe(`gatsby-source-drupal`, () => {
           })
         })
         it(`Creates node`, () => {
-          expect(nodes[createNodeId(`article-4`)]).toBeDefined()
-          expect(nodes[createNodeId(`article-4`)].title).toBe(`Article #4`)
+          expect(nodes[createNodeId(`und.article-4`)]).toBeDefined()
+          expect(nodes[createNodeId(`und.article-4`)].title).toBe(`Article #4`)
         })
 
         it(`Adds back references to referenced nodes`, () => {
           expect(
-            nodes[createNodeId(`tag-1`)].relationships[`node__article___NODE`]
-          ).toContain(createNodeId(`article-4`))
+            nodes[createNodeId(`und.tag-1`)].relationships[
+              `node__article___NODE`
+            ]
+          ).toContain(createNodeId(`und.article-4`))
         })
       })
     })
@@ -359,13 +376,13 @@ describe(`gatsby-source-drupal`, () => {
     const disallowedLinkTypes = [`self`, `describedby`, `taxonomy_term--tags`]
     await sourceNodes(args, { baseUrl, disallowedLinkTypes })
     expect(Object.keys(nodes).length).not.toEqual(0)
-    expect(nodes[createNodeId(`file-1`)]).toBeDefined()
-    expect(nodes[createNodeId(`file-2`)]).toBeDefined()
-    expect(nodes[createNodeId(`tag-1`)]).toBeUndefined()
-    expect(nodes[createNodeId(`tag-2`)]).toBeUndefined()
-    expect(nodes[createNodeId(`article-1`)]).toBeDefined()
-    expect(nodes[createNodeId(`article-2`)]).toBeDefined()
-    expect(nodes[createNodeId(`article-3`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.file-1`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.file-2`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.tag-1`)]).toBeUndefined()
+    expect(nodes[createNodeId(`und.tag-2`)]).toBeUndefined()
+    expect(nodes[createNodeId(`und.article-1`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.article-2`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.article-3`)]).toBeDefined()
   })
 
   it(`Verify JSON:API includes relationships`, async () => {
@@ -385,20 +402,20 @@ describe(`gatsby-source-drupal`, () => {
       entityReferenceRevisions,
     })
     expect(Object.keys(nodes).length).not.toEqual(0)
-    expect(nodes[createNodeId(`tag-1`)]).toBeUndefined()
-    expect(nodes[createNodeId(`tag-2`)]).toBeUndefined()
-    expect(nodes[createNodeId(`tag-3`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.tag-1`)]).toBeUndefined()
+    expect(nodes[createNodeId(`und.tag-2`)]).toBeUndefined()
+    expect(nodes[createNodeId(`und.tag-3`)]).toBeDefined()
     const paragraphForwardRevisionId = createNodeId(
-      `08d07c95-26ab-46b8-a56d-0a55567b2e31.4`
+      `und.08d07c95-26ab-46b8-a56d-0a55567b2e31.4`
     )
     const paragraphDraft = nodes[paragraphForwardRevisionId]
     expect(paragraphDraft).toBeDefined()
     expect(
-      nodes[createNodeId(`08d07c95-26ab-46b8-a56d-0a55567b2e31.3`)]
+      nodes[createNodeId(`und.08d07c95-26ab-46b8-a56d-0a55567b2e31.3`)]
     ).toBeDefined()
-    expect(nodes[createNodeId(`tag-3`)]).toBeDefined()
+    expect(nodes[createNodeId(`und.tag-3`)]).toBeDefined()
 
-    const article = nodes[createNodeId(`article-5`)]
+    const article = nodes[createNodeId(`und.article-5`)]
     expect(article).toBeDefined()
     const paragraphRelationships = article.relationships[`content___NODE`]
     expect(paragraphRelationships).toContain(paragraphForwardRevisionId)
@@ -418,24 +435,26 @@ describe(`gatsby-source-drupal`, () => {
         await sourceNodes(args, { baseUrl, fastBuilds })
       })
       it(`Attributes`, () => {
-        expect(nodes[createNodeId(`article-3`)].title).toBe(`Article #3`)
+        expect(nodes[createNodeId(`und.article-3`)].title).toBe(`Article #3`)
       })
       it(`Relationships`, () => {
-        expect(nodes[createNodeId(`article-3`)].relationships).toEqual({
-          field_main_image___NODE: createNodeId(`file-1`),
-          field_tags___NODE: [createNodeId(`tag-1`)],
+        expect(nodes[createNodeId(`und.article-3`)].relationships).toEqual({
+          field_main_image___NODE: createNodeId(`und.file-1`),
+          field_tags___NODE: [createNodeId(`und.tag-1`)],
         })
       })
       it(`Back references`, () => {
         expect(
-          nodes[createNodeId(`file-1`)].relationships[`node__article___NODE`]
-        ).toContain(createNodeId(`article-3`))
+          nodes[createNodeId(`und.file-1`)].relationships[
+            `node__article___NODE`
+          ]
+        ).toContain(createNodeId(`und.article-3`))
         expect(
-          nodes[createNodeId(`tag-1`)].relationships[`node__article___NODE`]
-        ).toContain(createNodeId(`article-3`))
+          nodes[createNodeId(`und.tag-1`)].relationships[`node__article___NODE`]
+        ).toContain(createNodeId(`und.article-3`))
         expect(
-          nodes[createNodeId(`tag-2`)].relationships[`node__article___NODE`]
-        ).not.toContain(createNodeId(`article-3`))
+          nodes[createNodeId(`und.tag-2`)].relationships[`node__article___NODE`]
+        ).not.toContain(createNodeId(`und.article-3`))
       })
     })
 
@@ -456,29 +475,31 @@ describe(`gatsby-source-drupal`, () => {
         await sourceNodes(args, { baseUrl, fastBuilds })
       })
       it(`Attributes`, () => {
-        expect(nodes[createNodeId(`article-3`)].title).toBe(
+        expect(nodes[createNodeId(`und.article-3`)].title).toBe(
           `Article #3 - Synced`
         )
       })
       it(`Relationships`, () => {
         // removed `field_main_image`, changed `field_tags`
-        expect(nodes[createNodeId(`article-3`)].relationships).toEqual({
-          field_tags___NODE: [createNodeId(`tag-2`)],
+        expect(nodes[createNodeId(`und.article-3`)].relationships).toEqual({
+          field_tags___NODE: [createNodeId(`und.tag-2`)],
         })
       })
       it(`Back references`, () => {
         // removed `field_main_image`, `file-1` no longer has back reference to `article-3`
         expect(
-          nodes[createNodeId(`file-1`)].relationships[`node__article___NODE`]
-        ).not.toContain(createNodeId(`article-3`))
+          nodes[createNodeId(`und.file-1`)].relationships[
+            `node__article___NODE`
+          ]
+        ).not.toContain(createNodeId(`und.article-3`))
         // changed `field_tags`, `tag-1` no longer has back reference to `article-3`
         expect(
-          nodes[createNodeId(`tag-1`)].relationships[`node__article___NODE`]
-        ).not.toContain(createNodeId(`article-3`))
+          nodes[createNodeId(`und.tag-1`)].relationships[`node__article___NODE`]
+        ).not.toContain(createNodeId(`und.article-3`))
         // changed `field_tags`, `tag-2` now has back reference to `article-3`
         expect(
-          nodes[createNodeId(`tag-2`)].relationships[`node__article___NODE`]
-        ).toContain(createNodeId(`article-3`))
+          nodes[createNodeId(`und.tag-2`)].relationships[`node__article___NODE`]
+        ).toContain(createNodeId(`und.article-3`))
       })
     })
   })
