@@ -1,5 +1,8 @@
 const axios = require(`axios`)
 const _ = require(`lodash`)
+const urlJoin = require(`url-join`)
+
+const { setOptions, getOptions } = require(`./plugin-options`)
 
 const {
   nodeFromData,
@@ -27,6 +30,10 @@ function gracefullyRethrow(activity, error) {
   }
 }
 
+exports.onPreBootstrap = (_, pluginOptions) => {
+  setOptions(pluginOptions)
+}
+
 exports.sourceNodes = async (
   {
     actions,
@@ -43,18 +50,23 @@ exports.sourceNodes = async (
   },
   pluginOptions
 ) => {
-  let {
+  const {
     baseUrl,
-    apiBase,
+    apiBase = `jsonapi`,
     basicAuth,
     filters,
     headers,
     params,
-    concurrentFileRequests,
-    disallowedLinkTypes,
-    skipFileDownloads,
-    fastBuilds,
+    concurrentFileRequests = 20,
+    disallowedLinkTypes = [`self`, `describedby`],
+    skipFileDownloads = false,
+    fastBuilds = false,
     entityReferenceRevisions = [],
+    languageConfig = {
+      defaultLanguage: `und`,
+      enabledLanguages: [`und`],
+      translatableEntities: [],
+    },
   } = pluginOptions
   const { createNode, setPluginStatus, touchNode } = actions
 
@@ -100,6 +112,7 @@ exports.sourceNodes = async (
             getNode,
             reporter,
             store,
+            languageConfig,
           },
           pluginOptions
         )
@@ -112,7 +125,6 @@ exports.sourceNodes = async (
     return
   }
 
-  fastBuilds = fastBuilds || false
   if (fastBuilds) {
     const lastFetched =
       store.getState().status.plugins?.[`gatsby-source-drupal`]?.lastFetched ??
@@ -128,7 +140,7 @@ exports.sourceNodes = async (
     try {
       // Hit fastbuilds endpoint with the lastFetched date.
       const data = await axios.get(
-        `${baseUrl}/gatsby-fastbuilds/sync/${lastFetched}`,
+        urlJoin(baseUrl, `gatsby-fastbuilds/sync/`, lastFetched.toString()),
         {
           auth: basicAuth,
           headers,
@@ -159,6 +171,9 @@ exports.sourceNodes = async (
                   createNodeIdWithVersion(
                     nodeSyncData.id,
                     nodeSyncData.type,
+                    getOptions().languageConfig
+                      ? nodeSyncData.attributes.langcode
+                      : `und`,
                     nodeSyncData.attributes?.drupal_internal__revision_id,
                     entityReferenceRevisions
                   )
@@ -185,6 +200,7 @@ exports.sourceNodes = async (
                   getNode,
                   reporter,
                   store,
+                  languageConfig,
                 },
                 pluginOptions
               )
@@ -210,18 +226,6 @@ exports.sourceNodes = async (
     `Fetch all data from Drupal`
   )
 
-  // Default apiBase to `jsonapi`
-  apiBase = apiBase || `jsonapi`
-
-  // Default disallowedLinkTypes to self, describedby.
-  disallowedLinkTypes = disallowedLinkTypes || [`self`, `describedby`]
-
-  // Default concurrentFileRequests to `20`
-  concurrentFileRequests = concurrentFileRequests || 20
-
-  // Default skipFileDownloads to false.
-  skipFileDownloads = skipFileDownloads || false
-
   // Fetch articles.
   reporter.info(`Starting to fetch all data from Drupal`)
 
@@ -229,7 +233,7 @@ exports.sourceNodes = async (
 
   let allData
   try {
-    const data = await axios.get(`${baseUrl}/${apiBase}`, {
+    const data = await axios.get(urlJoin(baseUrl, apiBase), {
       auth: basicAuth,
       headers,
       params,
@@ -239,6 +243,12 @@ exports.sourceNodes = async (
         if (disallowedLinkTypes.includes(type)) return
         if (!url) return
         if (!type) return
+
+        // Lookup this type in our list of language alterable entities.
+        const isTranslatable = languageConfig.translatableEntities.some(
+          entityType => entityType === type
+        )
+
         const getNext = async (url, data = []) => {
           if (typeof url === `object`) {
             // url can be string or object containing href field
@@ -249,7 +259,17 @@ exports.sourceNodes = async (
             // See https://www.drupal.org/docs/8/modules/jsonapi/filtering
             if (typeof filters === `object`) {
               if (filters.hasOwnProperty(type)) {
-                url = url + `?${filters[type]}`
+                url = new URL(url)
+                const filterParams = new URLSearchParams(filters[type])
+                const filterKeys = Array.from(filterParams.keys())
+                filterKeys.forEach(filterKey => {
+                  // Only add filter params to url if it has not already been
+                  // added.
+                  if (!url.searchParams.has(filterKey)) {
+                    url.searchParams.set(filterKey, filterParams.get(filterKey))
+                  }
+                })
+                url = url.toString()
               }
             }
           }
@@ -286,7 +306,34 @@ exports.sourceNodes = async (
           return data
         }
 
-        const data = await getNext(url)
+        let data = []
+        if (isTranslatable === false) {
+          data = await getNext(url)
+        } else {
+          for (let i = 0; i < languageConfig.enabledLanguages.length; i++) {
+            let currentLanguage = languageConfig.enabledLanguages[i]
+            const urlPath = url.href.split(`${apiBase}/`).pop()
+            const baseUrlWithoutTrailingSlash = baseUrl.replace(/\/$/, ``)
+            // The default language's JSON API is at the root.
+            if (
+              currentLanguage === getOptions().languageConfig.defaultLanguage ||
+              baseUrlWithoutTrailingSlash.slice(-currentLanguage.length) ==
+                currentLanguage
+            ) {
+              currentLanguage = ``
+            }
+
+            const joinedUrl = urlJoin(
+              baseUrlWithoutTrailingSlash,
+              currentLanguage,
+              apiBase,
+              urlPath
+            )
+            const dataForLanguage = await getNext(joinedUrl)
+
+            data = data.concat(dataForLanguage)
+          }
+        }
 
         const result = {
           type,
@@ -428,3 +475,37 @@ exports.onCreateDevServer = (
     }
   )
 }
+
+exports.pluginOptionsSchema = ({ Joi }) =>
+  Joi.object({
+    baseUrl: Joi.string()
+      .required()
+      .description(`The URL to root of your Drupal instance`),
+    apiBase: Joi.string().description(
+      `The path to the root of the JSONAPI — defaults to "jsonapi"`
+    ),
+    basicAuth: Joi.object({
+      username: Joi.string(),
+      password: Joi.string(),
+    }).description(`Enables basicAuth`),
+    filters: Joi.object().description(
+      `Pass filters to the JSON API for specific collections`
+    ),
+    headers: Joi.object().description(
+      `Set request headers for requests to the JSON API`
+    ),
+    params: Joi.object().description(`Append optional GET params to requests`),
+    concurrentFileRequests: Joi.number().integer().default(20).min(1),
+    disallowedLinkTypes: Joi.array().items(Joi.string()),
+    skipFileDownloads: Joi.boolean(),
+    fastBuilds: Joi.boolean(),
+    entityReferenceRevisions: Joi.array().items(Joi.string()),
+    secret: Joi.string().description(
+      `an optional secret token for added security shared between your Drupal instance and Gatsby preview`
+    ),
+    languageConfig: Joi.object({
+      defaultLanguage: Joi.string().required(),
+      enabledLanguages: Joi.array().items(Joi.string()).required(),
+      translatableEntities: Joi.array().items(Joi.string()).required(),
+    }),
+  })
