@@ -1,6 +1,6 @@
 import { GatsbyNodeApiHelpers } from "~/utils/gatsby-types"
 import merge from "lodash/merge"
-import { createRemoteMediaItemNode } from "~/steps/source-nodes/create-nodes/create-remote-media-item-node"
+import { createLocalFileNode } from "~/steps/source-nodes/create-nodes/create-local-file-node"
 import { menuBeforeChangeNode } from "~/steps/source-nodes/before-change-node/menu"
 import { cloneDeep } from "lodash"
 import { inPreviewMode } from "~/steps/preview"
@@ -234,7 +234,7 @@ const defaultPluginOptions: IPluginOptions = {
           actionType === `CREATE` ||
           actionType === `UPDATE`
         ) {
-          const createdMediaItem = await createRemoteMediaItemNode({
+          const createdMediaItem = await createLocalFileNode({
             mediaItemNode: remoteNode,
             parentName: `Node action ${actionType}`,
           })
