@@ -175,24 +175,42 @@ apiRunnerAsync(`onClientEntry`).then(() => {
       }
     ).pop()
 
-    const App = () => <GatsbyRoot>{SiteRoot}</GatsbyRoot>
+    const App = function App() {
+      const onClientEntryRanRef = React.useRef(false)
+
+      React.useEffect(() => {
+        if (!onClientEntryRanRef.current) {
+          onClientEntryRanRef.current = true
+          performance.mark(`onInitialClientRender`)
+
+          apiRunner(`onInitialClientRender`)
+        }
+      }, [])
+
+      return <GatsbyRoot>{SiteRoot}</GatsbyRoot>
+    }
 
     const renderer = apiRunner(
       `replaceHydrateFunction`,
       undefined,
-      ReactDOM.hydrate
+      process.env.GATSBY_EXPERIMENTAL_CONCURRENT_FEATURES
+        ? ReactDOM.unstable_createRoot
+        : ReactDOM.hydrate
     )[0]
 
     domReady(() => {
-      renderer(
-        <App />,
+      const container =
         typeof window !== `undefined`
           ? document.getElementById(`___gatsby`)
-          : void 0,
-        () => {
-          apiRunner(`onInitialClientRender`)
-        }
-      )
+          : null
+
+      if (renderer === ReactDOM.unstable_createRoot) {
+        renderer(container, {
+          hydrate: true,
+        }).render(<App />)
+      } else {
+        renderer(<App />, container)
+      }
     })
   })
 })
