@@ -2,6 +2,7 @@ import fs from "fs-extra"
 import glob from "glob"
 import path from "path"
 import webpack from "webpack"
+import _ from "lodash"
 import multer from "multer"
 import * as express from "express"
 import { urlResolve, getMatchPath } from "gatsby-core-utils"
@@ -18,15 +19,122 @@ const isProductionEnv = process.env.gatsby_executing_command !== `develop`
 
 interface IFunctionData {
   /** The route in the browser to access the function **/
-  apiRoute: string
+  functionRoute: string
+  /** The absolute path to the original function **/
+  originalAbsoluteFilePath: string
   /** The relative path to the original function **/
-  originalFilePath: string
+  originalRelativeFilePath: string
   /** The relative path to the compiled function (always ends with .js) **/
   relativeCompiledFilePath: string
   /** The absolute path to the compiled function (doesn't transfer across machines) **/
   absoluteCompiledFilePath: string
   /** The matchPath regex created by path-to-regexp. Only created if the function is dynamic. **/
   matchPath: string
+  /** The plugin that owns this function route **/
+  pluginName: string
+}
+
+interface IGlobPattern {
+  /** The plugin that owns this namespace **/
+  pluginName: string
+  /** The root path to the functions **/
+  rootPath: string
+  /** The glob pattern **/
+  globPattern: string
+}
+
+// During development, we lazily compile functions only when they're requested.
+// Here we keep track of which functions have been requested so are "active"
+const activeDevelopmentFunctions = new Set<IFunctionData>()
+let activeEntries = {}
+
+async function ensureFunctionIsCompiled(
+  functionObj: IFunctionData,
+  compiledFunctionsDir: string
+): any {
+  // stat the compiled function. If it's there, then return.
+  let compiledFileExists = false
+  try {
+    compiledFileExists = !!(await fs.stat(functionObj.absoluteCompiledFilePath))
+  } catch (e) {
+    // ignore
+  }
+  if (compiledFileExists) {
+    return
+  } else {
+    // Otherwise, restart webpack by touching the file and watch for the file to be
+    // compiled.
+    const time = new Date()
+    fs.utimesSync(functionObj.originalAbsoluteFilePath, time, time)
+    await new Promise(resolve => {
+      const watcher = chokidar
+        // Watch the root of the compiled function directory in .cache as chokidar
+        // can't watch files in directories that don't yet exist.
+        .watch(compiledFunctionsDir)
+        .on(`add`, async _path => {
+          if (_path === functionObj.absoluteCompiledFilePath) {
+            await watcher.close()
+
+            resolve(null)
+          }
+        })
+    })
+  }
+}
+
+// Create glob type w/ glob, plugin name, root path
+const createGlobArray = (siteDirectoryPath, plugins): Array<IGlobPattern> => {
+  const globs: Array<IGlobPattern> = []
+
+  // Add the default site src/api directory.
+  globs.push({
+    globPattern: `${siteDirectoryPath}/src/api/**/*.{js,ts}`,
+    rootPath: path.join(siteDirectoryPath, `src/api`),
+    pluginName: `default-site-plugin`,
+  })
+
+  // Add each plugin
+  plugins.forEach(plugin => {
+    // Ignore the "default" site plugin (aka the src tree) as we're
+    // already watching that.
+    if (plugin.name === `default-site-plugin`) {
+      return
+    }
+    // Ignore any plugins we include by default. In the very unlikely case
+    // we want to ship default functions, we'll special case add them. In the
+    // meantime, we'll avoid extra FS IO.
+    if (plugin.resolve.includes(`internal-plugin`)) {
+      return
+    }
+    if (plugin.resolve.includes(`gatsby-plugin-typescript`)) {
+      return
+    }
+    if (plugin.resolve.includes(`gatsby-plugin-page-creator`)) {
+      return
+    }
+
+    const glob = {
+      globPattern: `${plugin.resolve}/src/api/${plugin.name}/**/*.{js,ts}`,
+      rootPath: path.join(plugin.resolve, `src/api`),
+      pluginName: plugin.name,
+    } as IGlobPattern
+    globs.push(glob)
+  })
+
+  // Only return unique paths
+  return _.union(globs)
+}
+
+async function globAsync(pattern, options): Promise<Array<string>> {
+  return await new Promise((resolve, reject) => {
+    glob(pattern, options, (err, files) => {
+      if (err) {
+        reject(err)
+      } else {
+        resolve(files)
+      }
+    })
+  })
 }
 
 const createWebpackConfig = async ({
@@ -41,42 +149,51 @@ const createWebpackConfig = async ({
     `functions`
   )
 
-  const files = await new Promise((resolve, reject) => {
-    glob(`**/*.{js,ts}`, { cwd: functionsDirectory }, (err, files) => {
-      if (err) {
-        reject(err)
-      } else {
-        resolve(files)
-      }
+  const globs = createGlobArray(
+    siteDirectoryPath,
+    store.getState().flattenedPlugins
+  )
+
+  // Glob and return object with relative/absolute paths + which plugin
+  // they belong to.
+  const allFunctions = await Promise.all(
+    globs.map(async glob => {
+      const knownFunctions: Array<IFunctionData> = []
+      const files = await globAsync(glob.globPattern)
+      files.map(file => {
+        const originalAbsoluteFilePath = file
+        const originalRelativeFilePath = path.relative(glob.rootPath, file)
+
+        const { dir, name } = path.parse(originalRelativeFilePath)
+        // Ignore the original extension as all compiled functions now end with js.
+        const compiledFunctionName = path.join(dir, name + `.js`)
+        const compiledPath = path.join(
+          compiledFunctionsDir,
+          compiledFunctionName
+        )
+        const finalName = urlResolve(dir, name === `index` ? `` : name)
+
+        knownFunctions.push({
+          functionRoute: finalName,
+          pluginName: glob.pluginName,
+          originalAbsoluteFilePath,
+          originalRelativeFilePath,
+          relativeCompiledFilePath: compiledFunctionName,
+          absoluteCompiledFilePath: compiledPath,
+          matchPath: getMatchPath(finalName),
+        })
+      })
+
+      return knownFunctions
     })
-  })
+  )
 
-  if (files?.length === 0) {
-    reporter.warn(
-      `No functions found in directory: ${path.relative(
-        siteDirectoryPath,
-        functionsDirectory
-      )}`
-    )
-  }
-
-  const knownFunctions: Array<IFunctionData> = []
-  knownFunctions.forEach(f => f.apiRoute)
-  files.map(file => {
-    const { dir, name } = path.parse(file)
-    // Ignore the original extension as all compiled functions now end with js.
-    const compiledFunctionName = path.join(dir, name + `.js`)
-    const compiledPath = path.join(compiledFunctionsDir, compiledFunctionName)
-    const finalName = urlResolve(dir, name === `index` ? `` : name)
-
-    knownFunctions.push({
-      apiRoute: finalName,
-      originalFilePath: file,
-      relativeCompiledFilePath: compiledFunctionName,
-      absoluteCompiledFilePath: compiledPath,
-      matchPath: getMatchPath(finalName),
-    })
-  })
+  // Combine functions by the route name so that functions in the default
+  // functions directory can override the plugin's implementations.
+  const knownFunctions = _.unionBy(
+    ...allFunctions,
+    func => func.functionRoute
+  ) as Array<IFunctionData>
 
   store.dispatch(internalActions.setFunctions(knownFunctions))
 
@@ -134,18 +251,25 @@ const createWebpackConfig = async ({
   )
 
   const entries = {}
-  knownFunctions.forEach(({ originalFilePath }) => {
-    const filePath = path.join(functionsDirectory, originalFilePath)
-
+  const functionsList = isProductionEnv
+    ? knownFunctions
+    : activeDevelopmentFunctions
+  functionsList.forEach(functionObj => {
     // Get path without the extension (as it could be ts or js)
-    const parsedFile = path.parse(originalFilePath)
+    const parsedFile = path.parse(functionObj.originalRelativeFilePath)
     const compiledNameWithoutExtension = path.join(
       parsedFile.dir,
       parsedFile.name
     )
 
-    entries[compiledNameWithoutExtension] = filePath
+    entries[compiledNameWithoutExtension] = functionObj.originalAbsoluteFilePath
   })
+
+  activeEntries = entries
+
+  const stage = isProductionEnv
+    ? `functions-production`
+    : `functions-development`
 
   const config = {
     entry: entries,
@@ -155,6 +279,28 @@ const createWebpackConfig = async ({
       libraryTarget: `commonjs2`,
     },
     target: `node`,
+
+    // Minification is expensive and not as helpful for serverless functions.
+    optimization: {
+      minimize: false,
+    },
+
+    // Resolve files ending with .ts and the default extensions of .js, .json, .wasm
+    resolve: {
+      extensions: [`.ts`, `...`],
+    },
+
+    // Have webpack save its cache to the .cache/webpack directory
+    cache: {
+      type: `filesystem`,
+      name: stage,
+      cacheLocation: path.join(
+        siteDirectoryPath,
+        `.cache`,
+        `webpack`,
+        `stage-` + stage
+      ),
+    },
 
     mode: isProductionEnv ? `production` : `development`,
     // watch: !isProductionEnv,
@@ -178,6 +324,7 @@ const createWebpackConfig = async ({
   return config
 }
 
+let isFirstBuild = true
 export async function onPreBootstrap({
   reporter,
   store,
@@ -204,6 +351,7 @@ export async function onPreBootstrap({
   )
 
   await fs.ensureDir(compiledFunctionsDir)
+  await fs.emptyDir(compiledFunctionsDir)
 
   try {
     // We do this ungainly thing as we need to make accessible
@@ -240,10 +388,14 @@ export async function onPreBootstrap({
 
         // Log success in dev
         if (!isProductionEnv) {
-          reporter.success(`Re-building functions`)
+          if (isFirstBuild) {
+            isFirstBuild = false
+          } else {
+            reporter.success(`Re-building functions`)
+          }
         }
 
-        return resolve()
+        return resolve(null)
       }
 
       if (isProductionEnv) {
@@ -252,15 +404,28 @@ export async function onPreBootstrap({
         // When in watch mode, you call things differently
         let compiler = webpack(config).watch({}, callback)
 
+        const globs = createGlobArray(
+          siteDirectoryPath,
+          store.getState().flattenedPlugins
+        )
+
         // Watch for env files to change and restart the webpack watcher.
         chokidar
           .watch(
-            [`${siteDirectoryPath}/.env*`, `${siteDirectoryPath}/src/api/**/*`],
+            [
+              `${siteDirectoryPath}/.env*`,
+              ...globs.map(glob => glob.globPattern),
+            ],
             { ignoreInitial: true }
           )
-          .on(`all`, (event, path) => {
-            // Ignore change events from the API directory
-            if (event === `change` && path.includes(`/src/api/`)) {
+          .on(`all`, async (event, path) => {
+            // Ignore change events from the API directory for functions we're
+            // already watching.
+            if (
+              event === `change` &&
+              Object.values(activeEntries).includes(path) &&
+              path.includes(`/src/api/`)
+            ) {
               return
             }
 
@@ -295,9 +460,19 @@ export async function onCreateDevServer({
 }: CreateDevServerArgs): Promise<void> {
   reporter.verbose(`Attaching functions to development server`)
 
+  const {
+    program: { directory: siteDirectoryPath },
+  } = store.getState()
+
+  const compiledFunctionsDir = path.join(
+    siteDirectoryPath,
+    `.cache`,
+    `functions`
+  )
+
   app.use(
     `/api/*`,
-    multer().none(),
+    multer().any(),
     express.urlencoded({ extended: true }),
     (req, res, next) => {
       const cookies = req.headers.cookie
@@ -322,7 +497,7 @@ export async function onCreateDevServer({
 
       // Check first for exact matches.
       let functionObj = functions.find(
-        ({ apiRoute }) => apiRoute === pathFragment
+        ({ functionRoute }) => functionRoute === pathFragment
       )
 
       if (!functionObj) {
@@ -351,7 +526,11 @@ export async function onCreateDevServer({
       }
 
       if (functionObj) {
-        reporter.verbose(`Running ${functionObj.apiRoute}`)
+        activeDevelopmentFunctions.add(functionObj)
+
+        await ensureFunctionIsCompiled(functionObj, compiledFunctionsDir)
+
+        reporter.verbose(`Running ${functionObj.functionRoute}`)
         const start = Date.now()
         const pathToFunction = functionObj.absoluteCompiledFilePath
 
@@ -363,20 +542,26 @@ export async function onCreateDevServer({
 
           await Promise.resolve(fnToExecute(req, res))
         } catch (e) {
+          // Override the default error with something more specific.
+          if (e.message.includes(`fnToExecute is not a function`)) {
+            e.message = `${functionObj.originalAbsoluteFilePath} does not export a function.`
+          }
           reporter.error(e)
           // Don't send the error if that would cause another error.
           if (!res.headersSent) {
             res
               .status(500)
               .send(
-                `Error when executing function "${functionObj.originalFilePath}": "${e.message}"`
+                `Error when executing function "${functionObj.originalAbsoluteFilePath}":<br /><br />${e.message}`
               )
           }
         }
 
         const end = Date.now()
         reporter.log(
-          `Executed function "/api/${functionObj.apiRoute}" in ${end - start}ms`
+          `Executed function "/api/${functionObj.functionRoute}" in ${
+            end - start
+          }ms`
         )
       } else {
         next()
